requirements.txt
Django==1.10.5
django-cors-headers==1.3.1
djangorestframework==3.5.3
ez-setup==0.9
olefile==0.43
Pillow==4.0.0
psycopg2==2.6.2
